fix: useLocalStorage return type
If there is no value in localStorage it will
not return `undefined`, it will return `null`.

Also change default return type to string, instead
of `undefined` since it just can't have that value.

Even if you try to explicitly set a localStorage value
to `undefined` it will convert it to the string "undefined".